adding mock helper for locationUtilities
